% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ns.R
\name{check_returns}
\alias{check_returns}
\title{Check for disaggregation error corrections}
\usage{
check_returns(returns)
}
\arguments{
\item{returns}{dataset.}
}
\value{
Message
}
\description{
runs check of compiled time series data to identify where disaggregation error updates have occured
}
\author{
Daniel Braby
}
