% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_returns.R
\name{fix_returns}
\alias{fix_returns}
\title{Fix the returns that do not pass the validation test.}
\usage{
fix_returns(raw_returns, nursery_names)
}
\arguments{
\item{returns}{The raw returns.}
}
\description{
Fixes the disaggregation errors.
}
\author{
Daniel Braby
}
