% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_tables.R
\name{format_ts_diag}
\alias{format_ts_diag}
\title{Format table with year based time series to new or existing Excel spreadsheet for diagnostics checks}
\usage{
format_ts_diag(
  data_in,
  wb_in,
  sheet_name,
  latest_year_stage = NULL,
  data_out_names = NULL,
  year_var,
  main_table_text = NULL,
  inc_perc_chg = FALSE,
  inc_perc_tot = FALSE,
  sheet_title = "",
  main_title = ""
)
}
\arguments{
\item{data_in}{Data table as data frames or tibbles. The first column needs to contain the year, and be called "year" in any case.}

\item{wb_in}{Workbook object to write to.}

\item{sheet_name}{Text name of sheet.}

\item{latest_year_stage}{String to add to sheet that indicates the stage of the latest year's processing. One of: eg (expert group), prov (provisional), qa (quality assured), rev (revised) or final.}

\item{data_out_names}{Vector of names for columns output to spreadsheet tables, length the same as the number of columns in 'data_in' and column order the same as for 'data_in'.}

\item{year_var}{Name of variable containing year in data frame or tibble.}

\item{main_table_text}{Vector containing text to display in the rows below the main table. Each element of the vector is added on a new row.}

\item{inc_perc_chg}{Boolean, TRUE includes table of percentage change over last five years of 'data_in' below main table. Default is FALSE.}

\item{inc_perc_tot}{Boolean, TRUE includes table of percentage of total over last five years of 'data_in' below main table. Default is FALSE.}

\item{sheet_title}{Text title of sheet, to appear in first row and column.}

\item{main_title}{Text title of main data table, to appear above the table.}
}
\value{
out_wb - Workbook object
}
\description{
Format table with year based time series to new or existing Excel spreadsheet for diagnostics checks. There are some
custom formatting options available. See https://ycphs.github.io/openxlsx/reference/index.html for
further spreadsheet customisation options.
}
\author{
Kirsten Piller
}
