% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planting_year.R
\name{in_planting_year}
\alias{in_planting_year}
\title{Date to planting year interval.}
\usage{
in_planting_year(date)
}
\arguments{
\item{date}{Date to get the planting year from.}
}
\value{
Planting year interval.
}
\description{
Planting years run from 1st October to 30th September the following year.
}
\author{
Daniel Braby
}
