% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{names_diff}
\alias{names_diff}
\title{Recursively match file paths under directory.}
\usage{
names_diff(new, old)
}
\arguments{
\item{dir_path}{Directory path.}

\item{regex}{Matching criteria.}
}
\value{
File paths.
}
\description{
Return paths to files matching specified criteria.
}
\author{
Daniel Braby
}
