% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_pub.R
\name{ns_figures}
\alias{ns_figures}
\title{Create list of NS publication ready figures}
\usage{
ns_figures(returns, latest_year)
}
\arguments{
\item{returns}{compiled data}

\item{latest_year}{latest year of data}
}
\value{
List containing publication ready figures
}
\description{
Create list of NS publication ready figures
}
