% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nurserysurveyrap-package.R
\docType{package}
\name{nurserysurveyrap-package}
\alias{nurserysurveyrap}
\alias{nurserysurveyrap-package}
\title{nurserysurveyrap: nurserysurveyrap}
\description{
Read, analyse and report for our Nursery Survey.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/forestresearch/nurserysurveyrap.com}
  \item Report bugs at \url{http://github.com/forestresearch/nurseysurveyrap/issues.com}
}

}
\author{
\strong{Maintainer}: Dan Braby \email{daniel.braby@forestresearch.gov.uk}

}
\keyword{internal}
