% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_tables.R
\name{output_data_diag_xlsx}
\alias{output_data_diag_xlsx}
\title{Add data to new or existing Excel spreadsheet for diagnostics checks}
\usage{
output_data_diag_xlsx(
  data_in,
  sheet_name = NULL,
  out_path,
  out_name,
  latest_year,
  latest_year_stage = NULL,
  pub_date = NULL,
  next_update = NULL,
  data_out_names = NULL,
  year_var = NULL,
  main_table_text = NULL,
  diag_inc_perc_chg = FALSE,
  diag_inc_perc_tot = FALSE,
  replace_wb = FALSE,
  replace_sheet = FALSE,
  sheet_title = NULL,
  main_title = NULL
)
}
\arguments{
\item{data_in}{A data frame or tibble, or list of them for "pub_data". The first column needs to contain the year, and be called "year" in any case. If out_type = "pub", to work with a11ytables the list must contain first three data frames or tibbles; the first element for the cover sheet, second element for the contents sheet and third for the notes sheet. Then followed by any tables of data.}

\item{sheet_name}{Text name of sheet or, if out_type = "pub_cover", a vector containing two names for the Cover and Notes sheets respectively.}

\item{out_path}{Path to output Excel spreadsheet to.}

\item{out_name}{Name of output Excel spreadsheet.}

\item{latest_year}{Latest year of data.}

\item{latest_year_stage}{String to add to sheet that indicates the stage of the latest year's processing. One of: eg (expert group), prov (provisional), qa (quality assured), rev (revised) or final.}

\item{pub_date}{Publication date, string. Date format example: 6 October 2030. Default is NULL.}

\item{next_update}{Next update date, string. Date format example: 6 October 2030. Default is NULL.}

\item{data_out_names}{Vector of names for columns output to spreadsheet tables, or list of them for "pub_data" if data_in is a list. Length the same as the number of columns in 'data_in' and column order the same as for 'data_in'.}

\item{year_var}{Name of variable containing year in data frame or tibble.}

\item{main_table_text}{Vector containing text to display in the rows below the main table. Each element of the vector is added on a new row.}

\item{diag_inc_perc_chg}{If out_type = "diag". Boolean, TRUE includes table of percentage change over last five years of 'data_in' below main table. Default is FALSE.}

\item{diag_inc_perc_tot}{If out_type = "diag". Boolean, TRUE includes table of percentage of total over last five years of 'data_in' below main table. Default is FALSE.}

\item{replace_wb}{Boolean. If workbook already exists, TRUE replaces the workbook, FALSE will load the existing workbook. Default is FALSE.}

\item{replace_sheet}{Boolean. If workbook already exists, TRUE replaces the sheet, FALSE will stop the function. Default is FALSE.}

\item{sheet_title}{Text title of sheet, to appear in first row and column.}

\item{main_title}{Text title of main data table, to appear above the table.}
}
\description{
Add data to a new or existing Excel spreadsheet:
* a data frame or tibble containing a year based time series to new or existing Excel spreadsheet for diagnostics checks.

Diagnostics will have a simpler format designed for interaction and options to add summary statistics. Publication will have a publication ready format.
}
\author{
Kirsten Piller
}
