% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_tables.R
\name{output_data_xlsx_check}
\alias{output_data_xlsx_check}
\title{Checking file status prior to adding data to new or existing Excel spreadsheet for diagnostics checks or publication}
\usage{
output_data_xlsx_check(
  out_type,
  out_path,
  out_name,
  replace_wb = FALSE,
  replace_sheet = NULL
)
}
\arguments{
\item{out_type}{One of "pub" or "diag". "pub" uses a11ytables so different checks.}

\item{out_path}{Path to output Excel spreadsheet to.}

\item{out_name}{Name of output Excel spreadsheet.}

\item{replace_wb}{Boolean. If workbook already exists, TRUE replaces the workbook, FALSE will load the existing workbook. Default is FALSE.}

\item{replace_sheet}{Boolean. If workbook already exists, TRUE replaces the sheet, FALSE will stop the function. Default is NULL.}
}
\description{
Checking file status prior to adding data to new or existing Excel spreadsheet for diagnostics checks or publication. Will check for existence of file first.
}
\author{
Kirsten Piller
}
