% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_tables.R
\name{output_pub_tab}
\alias{output_pub_tab}
\title{Output publication tables to publication ready Excel spreadsheet}
\usage{
output_pub_tab(
  in_data,
  latest_year,
  pub_date,
  pub_name = NULL,
  which_chap = NULL,
  next_update,
  out_path,
  out_name_wb,
  replace_wb = FALSE,
  rnd_no_dec = NULL,
  cont_hyperlink = FALSE,
  a11y_obj = NULL
)
}
\arguments{
\item{in_data}{List of input tibbles or data frames for the a11ytable tables.}

\item{latest_year}{Latest year of data.}

\item{pub_date}{Publication date, string. Date format example: 6 October 2030.}

\item{pub_name}{Name of publication, which will determine the formatting. One of either "ukwpt" (UK Wood production and Trade) or "fs" (Forestry Statistics).}

\item{which_chap}{Number of the chapter to run.}

\item{next_update}{Next update date, string. Date format example: 6 October 2030.}

\item{out_path}{Path to output folder.}

\item{out_name_wb}{Name for output .xlsx workbook.}

\item{replace_wb}{Boolean. If workbook already exists, TRUE replaces the workbook, FALSE will load the existing workbook. Default is FALSE.}

\item{rnd_no_dec}{Vector containing TRUE or FALSE to indicate whether the values in that table of 'in_data' should be rounded to 0 decimal places, which the majority of tables usually are. Default is TRUE for all. Vector length needs to be the same as the list of input tables, 'in_data'.}

\item{cont_hyperlink}{Boolean. If TRUE, replaces sheet names in Contents sheet with hyperlinks to those sheets. Default is FALSE.}

\item{a11y_obj}{Optional. Supply list of objects required to run a11ytables if not using Forestry Statistics or UKWPT specific function pub_a11y_prep. Default is NULL.}
}
\description{
Output publication tables to publication ready Excel spreadsheet.

'in_data' tables should be rounded to the required decimal places already.
}
\author{
Kirsten Piller
}
