% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planting_year.R
\name{planting_year}
\alias{planting_year}
\title{Get the planting year as text.}
\usage{
planting_year(year)
}
\arguments{
\item{year}{Year to convert to planting year interval.}
}
\value{
Planting year text.
}
\description{
The interval \strong{starts} in the year that is specified.
}
