% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_fmt.R
\name{pub_a11y_prep}
\alias{pub_a11y_prep}
\title{Prepare content and formatting lists for use with a11ytables for a publication Excel spreadsheet}
\usage{
pub_a11y_prep(pub_date, next_update)
}
\arguments{
\item{pub_date}{Publication date, string. Date format example: 6 October 2030.}

\item{next_update}{Next update date, string. Date format example: 6 October 2030.}
}
\value{
List of cover, contents and note sheets for input to a11ytables and other formatting info.
}
\description{
Prepare content and formatting lists for use with a11ytables for a publication Excel spreadsheet
}
