% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nursery_names.R
\name{read_nursery_names}
\alias{read_nursery_names}
\title{Read a CSV with nursery_names and nursery_ref columns.}
\usage{
read_nursery_names(path)
}
\arguments{
\item{path}{Path to the CSV file.}
}
\value{
Tibble with correct column types.
}
\description{
Read a CSV with nursery_names and nursery_ref columns.
}
\author{
Daniel Braby
}
