% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{round_fmt_rmd}
\alias{round_fmt_rmd}
\title{Rounding and formatting function for R markdown}
\usage{
round_fmt_rmd(x, digits)
}
\arguments{
\item{x}{Number or vector of numbers.}

\item{digits}{Number specifying how many digits to round to.}
}
\value{
Rounded number or vector of numbers
}
\description{
Rounding and formatting function for R markdown that displays the correct number of decimal places in a rendered R markdown document.

For example, '3.0' would display as '3.0', rather than '3'.
}
