% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{round_safe}
\alias{round_safe}
\title{Rounding function with correction to rounding with 0.5}
\usage{
round_safe(x, digits)
}
\arguments{
\item{x}{Number or vector of numbers.}

\item{digits}{Number specifying how many digits to round to.}
}
\value{
Rounded number or vector of numbers
}
\description{
https://stackoverflow.com/questions/12688717/round-up-from-5

"Note that for rounding off a 5, the IEC 60559 standard is expected to be used, ‘go to the even digit’. Therefore round(0.5) is 0 and round(-1.5) is -2. However, this is dependent on OS services and on representation error (since e.g. 0.15 is not represented exactly, the rounding rule applies to the represented number and not to the printed number, and so round(0.15, 1) could be either 0.1 or 0.2)."
}
