% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planting_year.R
\name{to_planting_year}
\alias{to_planting_year}
\title{Get the planting year interval that is represented by a given year.}
\usage{
to_planting_year(year)
}
\arguments{
\item{year}{Year to convert to planting year interval.}
}
\value{
Planting year interval.
}
\description{
The interval \strong{starts} in the year that is specified.
}
