% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ns.R
\name{with_pivot}
\alias{with_pivot}
\title{Perform pivots}
\usage{
with_pivot(x, names_from, values_from, action)
}
\arguments{
\item{dir_path}{Directory path.}

\item{regex}{Matching criteria.}
}
\value{
File paths.
}
\description{
Bespoke pivot longer for Nursery Survey returns
}
\author{
Daniel Braby
}
