% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ns.R
\name{files_matching}
\alias{files_matching}
\title{Recursively match file paths under directory.}
\usage{
files_matching(dir_path, regex)
}
\arguments{
\item{dir_path}{Directory path.}

\item{regex}{Matching criteria.}
}
\value{
File paths.
}
\description{
Return paths to files matching specified criteria.
}
\author{
Daniel Braby
}
