% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_fmt.R
\name{ft_theme_pub}
\alias{ft_theme_pub}
\title{Format table with custom flextable theme for publication Word document}
\usage{
ft_theme_pub(in_table, round_it = TRUE, char_cols, num_cols)
}
\arguments{
\item{in_table}{Table (tibble or data frame) to be formatted}

\item{round_it}{Boolean, default is TRUE. TRUE will round numbers in table to 0 decimal places, FALSE will not round.}

\item{char_cols}{Vector containing index of column(s) that are character, can be length zero.}

\item{num_cols}{Vector containing index of column(s) that are numeric, can be length zero.}
}
\value{
Flextable
}
\description{
Format table with custom flextable theme for publication Word document
}
