% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_nursery.R
\name{output_nursery}
\alias{output_nursery}
\title{Output Nursery Survey}
\usage{
output_nursery(
  dir_path,
  hist_data,
  nursery_names,
  pub_year,
  latest_year,
  out_path,
  pub_date,
  next_update
)
}
\arguments{
\item{dir_path}{Path to folder containing all chapter input data files.}

\item{hist_data}{Name of previous historical time series of input data.}

\item{nursery_names}{List of corrected nursery names}

\item{latest_year}{String, year input data goes up to.}

\item{out_path}{Path to output folder.}

\item{pub_date}{Publication date, string. Date format example: 6 October 2030.}

\item{next_update}{Next update date, string. Date format example: 6 October 2030.}
}
\value{
tbd
}
\description{
Output Nursery Survey
}
