% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_returns.R
\name{read_return}
\alias{read_return}
\title{Read a single nursery survey return from a spreadsheet.}
\usage{
read_return(file_path)
}
\value{
Tibble with disaggregated country and genetically improved stock variables.
}
\description{
Read a single nursery survey return from a spreadsheet.
}
