% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_returns.R
\name{read_returns}
\alias{read_returns}
\title{Read all descendant spreadsheets in a folder and bind the reports together.}
\usage{
read_returns(dir_path)
}
\arguments{
\item{dir_path}{Path to search for spreadsheets.}
}
\value{
Combined tibble of reports.
}
\description{
Reads in all Nursery Survey returns in a specified spreadsheet and compiles into a single dataframe
}
\author{
Dan Braby
}
